replace codes of GraphicCoreLibrary/CameraManagement.h .cpp with HikSdk/video.h .cpp, delete HikSdk
