--- HikSdk/Wapper.h
@@ -1,4 +1,0 @@
-#ifndef  CUDA_Wapper_H
-#define CUDA_Wapper_H
-extern "C" bool YV12ToARGB(unsigned char* pYV12, unsigned char* pARGB, int width, int height);
-#endif
